Move sessionStore to tokenmanager.go
